requirements.txt
# client
sounddevice
numpy

# server
rpi_ws281x
adafruit-circuitpython-neopixel
